+++ apps/obelics/prj.conf
@@ -69,9 +69,6 @@ CONFIG_UDP_CLIENT=y
 # GNSS (M10)
 CONFIG_GNSS=y
 CONFIG_GNSS_U_BLOX_M10=y
-CONFIG_MODEM_STATS=y
-CONFIG_MODEM_MODULES=y
-CONFIG_MODEM_CHAT=y
 
 # I2C
 CONFIG_I2C=y
