+++ drivers/gnss/CMakeLists.txt
@@ -1,8 +1,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_library()
+# Parent directory - sources are in u_blox_m10 subdirectory
+zephyr_library_named(panoramix_gnss)
 
 # Add Zephyr's GNSS driver source directory for includes
 zephyr_include_directories(${ZEPHYR_BASE}/drivers/gnss)
 
-zephyr_library_sources_ifdef(CONFIG_GNSS_U_BLOX_M10 gnss_u_blox_m10.c gnss_u_blox_m10_i2c.c)
+add_subdirectory_ifdef(CONFIG_GNSS_U_BLOX_M10 u_blox_m10)
+
+# Allow empty in case no GNSS drivers are enabled
+set_property(TARGET panoramix_gnss PROPERTY ALLOW_EMPTY TRUE)
