+++ drivers/gnss/Kconfig
@@ -1,11 +1,4 @@
 # Copyright 2024
 # SPDX-License-Identifier: Apache-2.0
 
-config GNSS_U_BLOX_M10
-    bool "U-BLOX MAX-M10S GNSS Module (I2C)"
-    depends on GNSS
-    depends on DT_HAS_U_BLOX_M10_ENABLED
-    select GNSS_PARSE
-    select GNSS_NMEA0183_MATCH
-    help
-      Enable U-BLOX MAX-M10S GNSS driver over I2C.
+rsource "u_blox_m10/Kconfig"
