+++ drivers/gnss/u_blox_m10/gnss_u_blox_m10_i2c.h
