+++ dts/bindings/gnss/u-blox,m10.yaml
@@ -1,0 +1,18 @@
+# Copyright 2024
+# SPDX-License-Identifier: Apache-2.0
+
+description: U-BLOX MAX-M10S GNSS Module (I2C)
+
+compatible: "u-blox,m10"
+
+include: i2c-device.yaml
+
+properties:
+  label:
+    type: string
+    required: true
+
+  fix-rate-ms:
+    type: int
+    default: 1000
+    description: Fix rate in milliseconds
